+++ data/page_getEmbedList.js
@@ -1,23 +1,26 @@
 'use strict';
 
-let embed_selectors = ['iframe[src*="://www.dailymotion.com/embed/video/"]', 'iframe[src*="://www.hitbox.tv/"]', 'iframe[src*="://player.twitch.tv/?channel"]'];
-let embed_list = new Array();
-for(selector of embed_selectors){
+var embed_selectors = ['iframe[src*="://www.dailymotion.com/embed/video/"]', 'iframe[src*="://www.hitbox.tv/"]', 'iframe[src*="://player.twitch.tv/?channel"]'];
+
+var embed_list = new Array();
+for(let i in embed_selectors){
+	let selector = embed_selectors[i];
 	let result = document.querySelectorAll(selector);
 	if(result.length > 0){
-		for(node of result){
+		for(let j in result){
+			let node = result[j];
 			embed_list.push(node.src);
 		}
 	}
 }
 
-function sendDataToMain(){
+function sendDataToMain(portName){
 	this.port = chrome.extension.connect({name: portName});
 	this.sendData = function(id, data){
 		this.port.postMessage({"id": id, "data": data});
 	}
 }
-let my_port =	new sendDataToMain("Live_Streamer");
-let port = my_port.port;
+var my_port =	new sendDataToMain("Live_Streamer_Embed");
+var port = my_port.port;
 
 my_port.sendData("addStream", embed_list);